assistant/requirements.txt
fastapi==0.110.1
pydantic==2.6.4
pydantic-settings~=2.2.1
farm-haystack[elasticsearch]~=1.26.4